got sssp to work, I think

+++ Source/graph.h
@@ -12,5 +12,5 @@ spmat* read_graph(FILE *f);
 void write_graph(const spmat *graph, int dir, int weight, FILE *f);
 
 void bfs(const spmat *graph, long source, long *levels, long *parents);
-void sssp(const spmat *graph, long source, long *costs, long *parents);
+void sssp(const spmat *graph, long source, double *costs, long *parents);
 void apsp(const spmat *graph, double *dist);
